fix：修复SearchAnalysisService接口
